Create a BoundingVolumeSemantics namespace
